fix(db): align resource_data schema versions and test schemas
Record db_schema_version 3.6.0 in the fresh ClickHouse schema so it matches
the pom's db.version, correct the stale migration reference next to
resource_data, and make the test schemas use the same column types and
patient-inclusive sort key as the main schema.

+++ src/test/resources/schema.sql
@@ -720,7 +720,7 @@ CREATE TABLE resource_definition (
 CREATE TABLE resource_data (
     `RESOURCE_DATA_ID` Int64,
     `RESOURCE_ID`      String,
-    `CANCER_STUDY_ID`  Int64,
+    `CANCER_STUDY_ID`  Int32,
     `ENTITY_TYPE`      String,
     `PATIENT_ID`       Nullable(String),
     `SAMPLE_ID`        Nullable(String),
